File Size Optimization & Library Restructure
I put some files through an optimization tool to shrink the total file size down.
I Also changed the way libraries will be handled due to fast increasing file size with every program revision. This way, the large libraries are uploaded once and only new files are added each time.
For how to use the library folder, see the readme files in the repository root and in the program revisions.